refactor: split tests in separate files
